DEV-31 clock chart moved to Other section successfuly
